refactor(models): delete database.py shim; rewrite 8 callers to canonical paths
backend/models/database.py was a 86-line pure shim:
  - Re-exported Base + all ORM tables/enums from unified_models.
  - Aliased Prediction = MLPrediction.
  - Lazily proxied SessionLocal and get_db_session to backend.utils.database.

Eight production files imported from it (all scripts + 2 tests). Rewrites:
  scripts/validate_data.py              ORM imports -> unified_models
  scripts/test_sample_stocks.py         get_db_session -> backend.utils.database.get_db (aliased)
  scripts/testing/test_pipeline_components.py  same as above
  backend/tests/test_integration_comprehensive.py  Base -> unified_models
  backend/tests/test_database_integration.py       Base -> unified_models
  scripts/create_admin_user.py          User -> unified_models
  scripts/data/background_loader_enhanced.py  multi-symbol -> unified_models
  scripts/load_historical_data.py       multi-symbol -> unified_models

The Prediction alias and AlternativeData (never migrated, commented out in the
shim) had no live callers; both dropped without replacement. Grep for
"backend.models.database" against scripts/ and backend/ returns CLEAN.

Part of Tier B1 (docs/ANALYSIS_2026-05-14.md).
